solid/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
rdflib==7.0.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
cryptography==41.0.5
mimetypes-magic==0.1.0 